src/requirements.txt
requests
streamlit
python-dotenv
pandas
beautifulsoup4
lxml
pyyaml
aiohttp
aiodns
flask
playwright==1.40.0
pytest-playwright==0.4.3
setuptools-rust
psutil
pytest
fpdf2
SpeechRecognition
pyttsx3
openai-whisper
static-ffmpeg
playwright-stealth
fake-useragent
sqlalchemy
boto3
